Add in some alert tracking functionality.
Sometimes, metra posts an alert without updating any of the tracking
information for trains. This tries to capture that information and tell
you when you're specified line has an alert.

Unfortunately, this required that I add in another dependency on pup.
Also added a check to make sure everything is installed beforehand.